requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-groq
chromadb
sentence-transformers
python-dotenv
pandas
langchain-core
